Add cursor stack state

+++ src/main/java/com/apicatalog/cursor/jakarta/JakartaArrayItemCursor.java
@@ -17,7 +17,7 @@ public class JakartaArrayItemCursor extends JakartaValueCursor implements ArrayI
         if (!isArrayItem()) {
             throw new ClassCastException();    
         }        
-        return (Integer)cursor.index();
+        return cursor.index();
     }
 
     @Override
